Fix all compiler warnings (493 → 0)
Resolve all compiler warnings when building with -Wall -Wextra:

- Fix parentheses warnings (~375): Added parentheses to clarify operator
  precedence in bitwise operations across qrcode and decoder modules
- Fix sign-compare warnings (~62): Added appropriate casts to resolve
  signed/unsigned integer comparison issues
- Fix unused parameter warnings (~29): Marked unused parameters with
  (void) casts in stub implementations
- Fix override-init warnings (18): Added pragma to suppress intentional
  array initializer overwrites in symbol hash table
- Fix unterminated-string warnings (2): Increased array sizes to
  accommodate null terminators
- Fix fall-through warnings (2): Added fallthrough attributes to
  intentional switch case fall-throughs

All changes preserve existing functionality and maintain backward
compatibility. Build now completes with zero warnings and all tests pass.

+++ zbar/decoder/codabar.c
@@ -360,9 +360,9 @@ zbar_symbol_type_t _zbar_decode_codabar(zbar_decoder_t *dcode)
 	    goto reset;
 	}
 	n = codabar->character;
-	if (n < CFG(*codabar, ZBAR_CFG_MIN_LEN) ||
+	if ((int)n < CFG(*codabar, ZBAR_CFG_MIN_LEN) ||
 	    (CFG(*codabar, ZBAR_CFG_MAX_LEN) > 0 &&
-	     n > CFG(*codabar, ZBAR_CFG_MAX_LEN))) {
+	     (int)n > CFG(*codabar, ZBAR_CFG_MAX_LEN))) {
 	    dbprintf(2, " [invalid len]\n");
 	    goto reset;
 	}
